@@ -3,6 +3,7 @@ using CloudRepublic.BenchMark.Domain.Entities;
 using CloudRepublic.BenchMark.Domain.Enums;
 using System;
 using System.Collections.Generic;
+using CloudRepublic.BenchMark.Data;
 using Xunit;
 
 namespace CloudRepublic.BenchMark.API.Tests
@@ -19,7 +20,7 @@ namespace CloudRepublic.BenchMark.API.Tests
 
             var cloudProvider = CloudProvider.Azure;
             var hostingEnvironment = HostEnvironment.Windows;
-            var runtime = Runtime.Csharp;
+            var runtime = Runtime.FunctionsV4;
 
 
             var benchMarkResults = new List<BenchMarkResult>()
@@ -48,7 +49,7 @@ namespace CloudRepublic.BenchMark.API.Tests
             Assert.NotNull(benchMarkData);
             Assert.Equal(cloudProvider.ToString(), benchMarkData.CloudProvider);
             Assert.Equal(hostingEnvironment.ToString(), benchMarkData.HostingEnvironment);
-            Assert.Equal(runtime.ToString(), benchMarkData.Runtime);
+            Assert.Equal(runtime.GetName(), benchMarkData.Runtime);
 
 
         }
@@ -65,28 +66,28 @@ namespace CloudRepublic.BenchMark.API.Tests
                 {
                     CloudProvider =  CloudProvider.Firebase,
                     HostingEnvironment =  HostEnvironment.Linux,
-                    Runtime =  Runtime.Java,
+                    Language =  Language.Java,
                     CreatedAt = new DateTimeOffset(new DateTime(2020,1,1)),
                 },
                 new BenchMarkResult()
                 {
                     CloudProvider = CloudProvider.Azure, // different than the main
                     HostingEnvironment =  HostEnvironment.Windows,
-                    Runtime =  Runtime.Csharp,
+                    Language =  Language.Csharp,
                     CreatedAt = new DateTimeOffset(new DateTime(2020,1,1)),
                 },
                 new BenchMarkResult()
                 {
                     CloudProvider = CloudProvider.Azure, // different than the main
                     HostingEnvironment =  HostEnvironment.Windows,
-                    Runtime =  Runtime.Csharp,
+                    Language =  Language.Csharp,
                     CreatedAt = new DateTimeOffset(new DateTime(2020,1,1)),
                 },
                 new BenchMarkResult()
                 {
                     CloudProvider = CloudProvider.Azure, // different than the main
                     HostingEnvironment =  HostEnvironment.Windows,
-                    Runtime =  Runtime.Csharp,
+                    Language =  Language.Csharp,
                     CreatedAt = new DateTimeOffset(new DateTime(2020,1,1)),
                 },
             };
@@ -98,15 +99,12 @@ namespace CloudRepublic.BenchMark.API.Tests
             //  Act
 
             var benchMarkData = responseConverter.ConvertToBenchMarkData(benchMarkResults);
-
-
-
+            
             //  Assert
-
             Assert.NotNull(benchMarkData);
             Assert.Equal(CloudProvider.Firebase.ToString(), benchMarkData.CloudProvider);
             Assert.Equal(HostEnvironment.Linux.ToString(), benchMarkData.HostingEnvironment);
-            Assert.Equal(Runtime.Java.ToString(), benchMarkData.Runtime);
+            Assert.Equal(Language.Java.GetName(), benchMarkData.Language);
 
 
         }
@@ -123,27 +121,21 @@ namespace CloudRepublic.BenchMark.API.Tests
                 {
                     CloudProvider =  (CloudProvider)(-1) , // does not exist on enum
                     HostingEnvironment =(HostEnvironment)100, // does not exist on enum
-                    Runtime = (Runtime)789, // does not exist on enum
+                    Language = (Language)789, // does not exist on enum
                     CreatedAt = new DateTimeOffset(new DateTime(2020,1,1)),
                 },
             };
 
             var responseConverter = new ResponseConverterService();
 
-
-
             //  Act
-
             var benchMarkData = responseConverter.ConvertToBenchMarkData(benchMarkResults);
 
-
-
             //  Assert
-
             Assert.NotNull(benchMarkData);
-            Assert.Equal("-1", benchMarkData.CloudProvider);
-            Assert.Equal("100", benchMarkData.HostingEnvironment);
-            Assert.Equal("789", benchMarkData.Runtime);
+            Assert.Null(benchMarkData.CloudProvider);
+            Assert.Null(benchMarkData.HostingEnvironment);
+            Assert.Null(benchMarkData.Language);
 
 
         }
